full c implementation

+++ src/imaging/image.c
@@ -12,37 +12,37 @@ supports imgs of size max size 128x128
 
 // Image structures
 struct Image{
-	uint8_t size;
+	uint16_t size;
 	float **data;
 };
 
 struct CImage{
-	uint8_t size;
+	uint16_t size;
 	float **re;
 	float **im;
 };
 
 // allocate memory and zero init image
-void initImage(struct Image image, uint8_t size){
+void initImage(struct Image image, uint16_t size){
 	image.size = size;
 	image.data = malloc(size * sizeof(float *));
-	for(uint8_t row = 0; row < size; i++){
+	for(uint16_t row = 0; row < size; i++){
 		image.data[row] = malloc(size * sizeof(float));
-		for(uint8_t col = 0; col < image.size){
+		for(uint16_t col = 0; col < image.size){
 			image.data[row][col] = 0;
 		}
 	}
 }
 
 // allocate memory for complex img representation
-void initComplex(struct CImage image, uint8_t size){
+void initComplex(struct CImage image, uint16_t size){
 	image.size = size;
 	image.re = malloc(size * sizeof(float *));
 	image.im = malloc(size * sizeof(float *));
-	for(uint8_t row = 0; row < size; i++){
+	for(uint16_t row = 0; row < size; i++){
 		image.re[row] = malloc(size * sizeof(float));
 		image.im[row] = malloc(size * sizeof(float));
-		for(uint8_t col = 0; col < image.size){
+		for(uint16_t col = 0; col < image.size){
 			image.re[row][col] = 0;
 			image.im[row][col] = 0;
 		}
@@ -51,8 +51,8 @@ void initComplex(struct CImage image, uint8_t size){
 
 // zero out content of image
 void resetImage(struct Image image){
-	for(uint8_t row = 0; row < image.size){
-		for(uint8_t col = 0; col < image.size){
+	for(uint16_t row = 0; row < image.size){
+		for(uint16_t col = 0; col < image.size){
 			image.data[row][col] = 0;
 		}
 	}
@@ -60,8 +60,8 @@ void resetImage(struct Image image){
 
 // zero out content of image
 void resetImage(struct CImage image){
-	for(uint8_t row = 0; row < image.size){
-		for(uint8_t col = 0; col < image.size){
+	for(uint16_t row = 0; row < image.size){
+		for(uint16_t col = 0; col < image.size){
 			image.re[row][col] = 0;
 			image.im[row][col] = 0;
 		}
@@ -70,7 +70,7 @@ void resetImage(struct CImage image){
 
 // deallocate memory in image
 void deinitImage(struct Image image){
-	for(uint8_t row = 0; row < image.size){
+	for(uint16_t row = 0; row < image.size){
 		free(image.data[row]);
 	}
 	free(image.data);
@@ -79,7 +79,7 @@ void deinitImage(struct Image image){
 
 // deallocate memory for complex represenation
 void deinitImage(struct CImage image){
-	for(uint8_t row = 0; row < image.size){
+	for(uint16_t row = 0; row < image.size){
 		free(image.re[row]);
 		free(image.im[row]);
 	}
@@ -89,8 +89,8 @@ void deinitImage(struct CImage image){
 
 // load entire image into structure
 void loadImage(struct Image image, float *data_stream){
-	for(uint8_t row = 0; row < image.size){
-		for(uint8_t col = 0; col < image.size){
+	for(uint16_t row = 0; row < image.size){
+		for(uint16_t col = 0; col < image.size){
 			image.data[row][col] = data_stream[row + image.size * col];
 		}
 	}
@@ -98,23 +98,23 @@ void loadImage(struct Image image, float *data_stream){
 
 // load entire image into complex structure
 void loadComplexImage(struct CImage image, float *data_stream){
-	for(uint8_t row = 0; row < image.size){
-		for(uint8_t col = 0; col < image.size){
+	for(uint16_t row = 0; row < image.size){
+		for(uint16_t col = 0; col < image.size){
 			image.re[row][col] = data_stream[row + image.size * col];
 		}
 	}
 }
 
 // load a single row into img structure
-void loadRow(struct Image image, float *row_data, uint8_t index){
-	for(uint8_t col = 0; col < image.size; ++column){
+void loadRow(struct Image image, float *row_data, uint16_t index){
+	for(uint16_t col = 0; col < image.size; ++column){
 		image.data[index][column] = row_data[column];
 	}
 }
 
 // load a single row into complex img structure
-void loadComplexRow(struct CImage image, float *row_data, uint8_t index){
-	for(uint8_t col = 0; col < image.size; ++column){
+void loadComplexRow(struct CImage image, float *row_data, uint16_t index){
+	for(uint16_t col = 0; col < image.size; ++column){
 		image.re[index][column] = row_data[column];
 	}
 }
